+++ app/src/main/java/woowacourse/movie/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import woowacourse.movie.domain.Movie
 import java.time.LocalDate
 
 class MainActivity : AppCompatActivity() {
