+++ app/src/main/java/woowacourse/movie/MovieAdapter.kt
@@ -9,6 +9,7 @@ import android.widget.BaseAdapter
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import woowacourse.movie.domain.Movie
 import java.time.format.DateTimeFormatter
 
 class MovieAdapter(
